Home work. Kotlin lesson1

+++ app/src/main/java/com/example/lesson10/NoteListFragment.java
@@ -8,20 +8,22 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
+import android.widget.Toast;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.widget.PopupMenu;
 import androidx.core.content.res.ResourcesCompat;
+import androidx.fragment.app.DialogFragment;
 import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.DefaultItemAnimator;
 import androidx.recyclerview.widget.DividerItemDecoration;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import java.util.Calendar;
 import java.util.Objects;
 
-public class NoteListFragment extends Fragment{
+public class NoteListFragment extends Fragment {
 
     private RecyclerView recyclerView;
     private Context context;
@@ -42,14 +44,14 @@ public class NoteListFragment extends Fragment{
         initView(view);
         // создаем подключение и получаем данные. метод init ооповещает обозревателей, т.к метод
         // асинхронный
-        data = new NoteSourceFirebaseImp().init(noteSource ->adapter.notifyDataSetChanged());
+        data = new NoteSourceFirebaseImp().init(noteSource -> adapter.notifyDataSetChanged());
         adapter.setDataSource(data);
         initPopupMenu(view);
     }
 
     private void initView(View view) {
         TextView userName = view.findViewById(R.id.user_name);
-        userName.setText(User.nameUser);
+        userName.setText(User.INSTANCE.getNameUser());
         recyclerView = view.findViewById(R.id.recycler_view_notes);
         initRecyclerView();
     }
@@ -63,13 +65,16 @@ public class NoteListFragment extends Fragment{
             popup.show();
             popup.setOnMenuItemClickListener(item -> {
                 if (item.getItemId() == R.id.action_add) {
-                    Note newNote = new Note("", "",Calendar.getInstance().getTime());
-                    addUpdateNote(newNote, data.size());
+                    //2.* Переделайте ввод и редактирование данных через диалоговое окно.
+                    DialogFragment dlgBuilder = new NoteDialogCreateFragment(data.size());
+                    MainActivity activity = (MainActivity) context;
+                    Navigation navigation = activity.getNavigation();
+                    dlgBuilder.show(navigation.getFragmentManager(), "NOTE_CREATE");
                     return true;
                 } else if (item.getItemId() == R.id.action_clear) {
                     data.clearNoteData();
                     //Обновляет данные списка.
-                    data.init(noteSource ->adapter.notifyDataSetChanged());
+                    data.init(noteSource -> adapter.notifyDataSetChanged());
                     return true;
                 }
                 return true;
@@ -90,7 +95,7 @@ public class NoteListFragment extends Fragment{
         }
     }
 
-     private void initRecyclerView() {
+    private void initRecyclerView() {
 
         recyclerView.setHasFixedSize(true);
 
@@ -117,9 +122,34 @@ public class NoteListFragment extends Fragment{
                 ((Controller) requireActivity())
                         .openNoteScreen(adapter.getDataSource().getNoteData(position), position);
             } else if (itemId == adapter.CMD_DELETE) {
-                data.deleteNoteData(position);
-                adapter.notifyDataSetChanged();
+                // 1. Создайте диалоговое окно с предупреждением перед удалением данных.
+
+                // Создаём билдер и передаём контекст приложения
+                AlertDialog.Builder builder = new AlertDialog.Builder(context);
+                // В билдере указываем заголовок окна. Можно указывать как ресурс, так
+                // и строку
+                builder.setTitle(R.string.alert_dialog_delete)
+                        // Указываем сообщение в окне. Также есть вариант со строковым
+                        // параметром
+                        .setMessage(R.string.exclamation)
+                        // Из этого окна нельзя выйти кнопкой Back
+                        .setCancelable(false)
+                        // Устанавливаем отрицательную кнопку
+                        .setNegativeButton(R.string.no,
+                                // Ставим слушатель, будем обрабатывать нажатие
+                                (dialog, id) -> Toast.makeText(context, "Удаление отменено", Toast.LENGTH_SHORT).show())
+                        // Устанавливаем кнопку. Название кнопки также можно задавать
+                        // строкой
+                        .setPositiveButton(R.string.yes,
+                                // Ставим слушатель, будем обрабатывать нажатие
+                                (dialog, id) -> {
+                                    data.deleteNoteData(position);
+                                    adapter.notifyDataSetChanged();
+                                });
+                AlertDialog alert = builder.create();
+                alert.show();
             }
+
         });
     }
 
@@ -129,7 +159,7 @@ public class NoteListFragment extends Fragment{
             data.updateNoteData(position, newNote);
         } else data.addNoteData(newNote);
         //метод init ооповещает обозревателей
-        data.init(noteSource ->adapter.notifyDataSetChanged());
+        data.init(noteSource -> adapter.notifyDataSetChanged());
         //позицианируется на новой позиции
         recyclerView.scrollToPosition(position);
     }
